+++ client/components/ui/Header/Header.tsx
@@ -6,7 +6,7 @@ import { useRouter } from 'expo-router';
 import { useTheme } from '@/providers/themeProvider';
 import { View } from '../View';
 import { Text } from '../Text';
-import { Inline } from '../Layout';
+import { Inline } from '../Layout/Layout';
 import { TouchableOpacity } from 'react-native';
 import { HeaderProps, variantType } from './Header.interface';
