+++ client/components/ui/index.ts
@@ -4,5 +4,5 @@ export * from './View';
 export * from './Button/Button';
 export * from './Input/Input';
 export * from './Card/Card';
-export * from './Layout';
+export * from './Layout/Layout';
 export * from './Header/Header';
